% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-maps.R
\name{maplibre_h3Output}
\alias{maplibre_h3Output}
\title{Shiny output function for \code{maplibre_h3}}
\usage{
maplibre_h3Output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Widget dimensions (CSS string or pixels).}
}
\value{
A Shiny output object.
}
\description{
Shiny output function for \code{maplibre_h3}
}
