% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-maps.R
\name{renderMaplibre_h3}
\alias{renderMaplibre_h3}
\title{Shiny render function for \code{maplibre_h3}}
\usage{
renderMaplibre_h3(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a \code{maplibre_h3} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Logical. Is \code{expr} already quoted? Default: \code{FALSE}.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for \code{maplibre_h3}
}
